[feat] 24.03.22 약 상세정보 요약기능 추가

+++ backend/src/main/java/com/yoyak/yoyak/medicineDetail/domain/MedicineDetailRepository.java
@@ -3,7 +3,8 @@ package com.yoyak.yoyak.medicineDetail.domain;
 import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MedicineDetailRepository extends JpaRepository<MedicineDetail, Integer> {
+public interface MedicineDetailRepository extends JpaRepository<MedicineDetail, Long> {
 
     Optional<MedicineDetail> findBySeq(Long Seq);
+
 }
